The Dual
This is some test script I wrote for a game mechanic whilst learning Python. There's alot to improve, but I'm gonna list the improvements I want to make and the problems that need solving then move on with my module, rather than spend forever on this. I'll definitely return to it later, and use it in future games and mechanics

Changes and improvements:
1) Full spell checking of strings.
2) Find away to force the player to input a integer 1, 2 or 3, or return some kind of error.
3) Get the game to end immediately when a players HP reaches 0, probably will have to ditch the loop mechanic. 
4)Find a better way for the the appear. might add more pointless inputs to force the player to push enter,