Create The Dual
This script works to test my idea for a combat mechanic for future python games. I wrote it as part of my Learn Python The Hard Way Exercises, so I was focusing on using certain statements and commands, which might not have been the actual best choice. It has certain flaws, and bugs, but I'l list those here for when I want to improve it int he future, rather than try and sort them out now.

1. Spell check all the strings, and improve the language
2. Find a way to make the user input an integer 1, 2 or 3. it runs regardless as is.
3. When the Player or NPC HP hits zero, the loop will keep running, until it completes one full loop. I need to find a way for the loop to finish the moment a player hits 0 HP.